+++ src/utils/free/minishell.c
@@ -1,6 +1,20 @@
 #include "minishell.h"
 
 /**
+ * Frees any pointer and sets it to NULL.
+ *
+ * @param ptr The address of the pointer to free.
+ */
+void	ft_free_ptr(void **ptr)
+{
+	if (ptr && *ptr)
+	{
+		free(*ptr);
+		*ptr = NULL;
+	}
+}
+
+/**
  * Free minishell at any point during the program execution.
  *
  * @param minishell Struct containing global data on the program,
