+++ src/utils/string.c
@@ -1,41 +1,6 @@
 #include "minishell.h"
 
 /**
- * Checks if the character is a special character (|, <, >).
- * Used to differentiate between operators and words in tokenization.
- *
- * @param c The character to check
- * @return 1 if special char, 0 otherwise.
- */
-int	is_special_char(char c)
-{
-	return (c == '|' || c == '<' || c == '>');
-}
-
-/**
- * Checks if the character is a whitespace (space, tab, newline, etc.).
- *
- * @param c The character to check
- * @return 1 if it is a whitespace, 0 otherwise).
- */
-int	is_space_char(char c)
-{
-	return (c == ' ' || (c >= 9 && c <= 13));
-}
-
-/**
- * Checks if the character is a quote symbol.
- * Used to determine whether a word token is enclosed in quotes.
- *
- * @param c The character to check
- * @return 1 if quote char, 0 otherwise.
- */
-int	is_quote_char(char c)
-{
-	return (c == '\'' || c == '"');
-}
-
-/**
  * Skips all leading whitespace characters in the input string.
  * Advances the `i` index to the next non-whitespace character.
  *
@@ -74,7 +39,7 @@ char	*char_to_str(char c)
  * @param minishell Pointer to the minishell struct for cleanup on failure.
  * @return A pointer to a static string representing the integer.
  */
-char *int_to_str(int n)
+char	*int_to_str(int n)
 {
 	static char	buf[12];
 	char		*tmp;
