+++ test/bash_signals_recap
@@ -20,7 +20,7 @@
 
 ---
 
-# TODO
+# 
 
 - Ctrl-D : OK
 - Ctrl-C :
